The major point of this commit was to use the Api in the ui-router.
ExpenseController.java - Added request mappings so it's possible to post to /expenses. When dealing with a collection of resources posting a new resource to self makes sense.
ExpenseResourceLinksAssembler.java - Tried to create a templated link...
Api.js - I had to include the http client directly in the api module since requirejs is giving me trouble.
app.js - I added an abstract state where the Api is resolved for a given entry point. Subsequently this is used in the api.expenses state to retrieve the expenses.